Reorganized preprocessing and aggregation functions
